Sending to Python from jack works
Not sure if we are successfully receiving messages from Python to be sent out to
jack, e.g., we cannot get playback via Fluidsynth. But almost there.

Note: the out MessageQueue of midi_proc_jack.c is the in MessageQueue of
midi_proc_jack.py and vice versa for the output MessageQueue.